requirements.txt
databricks-sql-connector>=3.0.0
pyarrow>=14.0.0
pandas>=2.0.0
geopandas>=0.14.0
shapely>=2.0.0
networkx>=3.0
lxml>=5.0.0
python-dotenv>=1.0.0
tqdm>=4.0.0
